+++ app/controllers/Application.java
@@ -14,6 +14,7 @@ import java.util.List;
 import models.Account;
 import models.AccountRole;
 import models.survey.Survey;
+import models.survey.SurveyType;
 import play.CorePlugin;
 import play.cache.Cache;
 import play.i18n.Lang;
@@ -48,9 +49,12 @@ public class Application extends Controller {
     Account me = Security.connected();
     notFoundIfNull(me);
     
-    List<Survey> surveys = Survey.find("isDeleted=false and isPublished=true").fetch();
+    List<Survey> surveys = Survey.find("type=:type and isDeleted=false and isPublished=true")
+        .setParameter("type", SurveyType.REGULAR).fetch();
+    List<Survey> standards = Survey.find("type=:type and isDeleted=false and isPublished=true")
+        .setParameter("type", SurveyType.STANDARD).fetch();
 
-    render("application/dashboard.html", surveys);
+    render("application/dashboard.html", surveys, standards);
   }
 
   /**