+++ app/models/survey/SurveyType.java
@@ -1,0 +1,9 @@
+/**
+ * 
+ */
+package models.survey;
+
+public enum SurveyType {
+  REGULAR,
+  STANDARD
+}